refactor(kernel): restructure modules into apps/drivers, use resonance tracing
