+++ fullerene-kernel/src/apps/mod.rs
@@ -1,0 +1,2 @@
+pub mod app_runner;
+pub mod badapple;
