+++ fullerene-kernel/src/drivers/mod.rs
@@ -1,0 +1,3 @@
+pub mod ahci;
+pub mod nvme;
+pub mod virtio_gpu;
