+++ fullerene-kernel/src/init.rs
@@ -166,7 +166,7 @@ pub fn init_common(_physical_memory_offset: x86_64::VirtAddr) {
             Ok(())
         }),
         petroleum::init_step!("app_runner", || {
-            crate::app_runner::init();
+            crate::apps::app_runner::init();
             petroleum::serial::serial_log(format_args!("App runner initialised\n"));
             crate::boot_stage!(BootStage::AppRunnerReady);
             Ok(())
