+++ fullerene-kernel/src/main.rs
@@ -5,9 +5,6 @@
 extern crate alloc;
 
 // ---- Custom panic handler (replaces petroleum::define_panic_handler!) ----
-// We define our own so we can call klog::flush_to_vfs_safe() directly,
-// without going through the extern "Rust" indirection that would conflict
-// with petroleum's empty stub.
 #[cfg(all(any(target_os = "none", target_os = "uefi"), not(test)))]
 #[panic_handler]
 fn panic(info: &core::panic::PanicInfo) -> ! {
@@ -33,18 +30,19 @@ fn panic(info: &core::panic::PanicInfo) -> ! {
 petroleum::define_alloc_error_handler!();
 
 // Constants
-/// Physical address of the VGA text-mode buffer.
-/// Currently unused but kept as a reference for potential fallback debugging.
-#[allow(unused)]
 pub const VGA_BUFFER_ADDRESS: usize = 0xb8000;
 
 // Exported globals
 pub use heap::MEMORY_MAP;
 
 // Module declarations
-// pub mod ahci;  // disabled: unused on current hardware targets
-pub mod app_runner;
-pub mod badapple;
+// ── Applications (user-visible programs, demos) ────────────────────
+pub mod apps;
+
+// ── Drivers (storage, GPU, network) ───────────────────────────────
+pub mod drivers;
+
+// ── Kernel core ────────────────────────────────────────────────────
 pub mod boot;
 pub mod boot_stage;
 pub mod context_switch;
@@ -60,13 +58,12 @@ pub mod interrupts;
 pub mod klog;
 pub mod loader;
 pub mod memory_management;
-pub mod nvme;
 pub mod process;
 pub mod scheduler;
 pub mod shell;
 pub mod slab;
 pub mod syscall;
 pub mod task;
+// tracing.rs is now a thin re-export of resonance::tracing
 pub mod tracing;
 pub mod vfs;
-pub mod virtio_gpu;
