+++ resonance/Cargo.toml
@@ -5,3 +5,6 @@ edition.workspace = true
 authors.workspace = true
 
 [dependencies]
+x86_64 = { version = "0.15.2", default-features = false, features = [
+  "instructions",
+] }
