+++ resonance/src/lib.rs
@@ -7,9 +7,11 @@ mod event;
 mod handler;
 mod queue;
 mod source;
+pub mod tracing;
 
 pub use dispatcher::*;
 pub use event::*;
 pub use handler::*;
 pub use queue::*;
 pub use source::*;
+pub use tracing::TraceEvent;
